Refactor for better encapsulation
