+++ src/hooks/useModal.js
@@ -4,7 +4,6 @@ import logMessage from '../utilities/loggingConfig';
 const useModal = () => {
     const [isModalOpen, setIsModalOpen] = useState(false);
     const [modalParams, setModalParams] = useState({});
-    const [installPromptEvent, setInstallPromptEvent] = useState(null);
 
     const openModal = (params) => {
         setModalParams(params);
@@ -15,28 +14,11 @@ const useModal = () => {
         setIsModalOpen(false);
     };
 
-    const captureInstallPrompt = (event) => {
-        logMessage(`Install prompt captured: ${event.type}`);
-        event.preventDefault();
-        setInstallPromptEvent(event);
-    };
-
-    const showInstallPrompt = () => {
-        logMessage(`Showing install prompt: ${installPromptEvent.type}`);
-        if (installPromptEvent) {
-            installPromptEvent.prompt();
-            setInstallPromptEvent(null);
-        }
-    };
-
     return {
         isModalOpen,
         modalParams,
         openModal,
-        closeModal,
-        captureInstallPrompt,
-        showInstallPrompt,
-        installPromptEvent // Ensure this is exposed
+        closeModal
     };
 };
